requirements.txt
requests>=2.32.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
APScheduler>=3.10.0
lxml>=5.2.0
